// backend/src/shared/enums/error-code.ts
export enum ErrorCode {
  INTERNAL_SERVER_ERROR = "INTERNAL_SERVER_ERROR",
  INCORRECT_CREDENTIALS_ERROR = "INCORRECT_CREDENTIALS_ERROR",
  REFRESH_TOKEN_NOT_FOUND = "REFRESH_TOKEN_NOT_FOUND",
  CONFIRM_USER_TOKEN_NOT_FOUND = "CONFIRM_USER_TOKEN_NOT_FOUND",
  INVALID_TOKEN = "INVALID_TOKEN",
  EXPIRED_TOKEN = "EXPIRED_TOKEN",
  INVALID_CPF = "INVALID_CPF",
  INVALID_PHONE = "INVALID_PHONE",
  INVALID_EMAIL = "INVALID_EMAIL",
  INVALID_PASSWORD_HASH = "INVALID_PASSWORD_HASH",
  INVALID_COUNTRY = "INVALID_COUNTRY",
  USER_NOT_FOUND = "USER_NOT_FOUND",
  USER_ALREADY_EXISTS = "USER_ALREADY_EXISTS",
  USER_DEACTIVATED = "USER_DEACTIVATED",
  USER_NOT_CONFIRMED = "USER_NOT_CONFIRMED",
  USER_MISSING_PASSWORD = "USER_MISSING_PASSWORD",
  USER_ALREADY_CONFIRMED = "USER_ALREADY_CONFIRMED",
}